Move all lore item code to single file, delete sidebar
