+++ src/module/item/index.ts
@@ -8,11 +8,11 @@ export { ConditionPF2e } from "./condition/document.ts";
 export { EffectPF2e } from "./effect/document.ts";
 
 // Physical Items
-export { PhysicalItemPF2e } from "./physical/document.ts";
 export { ArmorPF2e } from "./armor/document.ts";
 export { BookPF2e } from "./book/document.ts";
 export { ContainerPF2e } from "./container/document.ts";
 export { EquipmentPF2e } from "./equipment/document.ts";
+export { PhysicalItemPF2e } from "./physical/document.ts";
 export { TreasurePF2e } from "./treasure/document.ts";
 export { WeaponPF2e } from "./weapon/document.ts";
 
@@ -24,13 +24,13 @@ export { ClassPF2e } from "./class/document.ts";
 
 // Others
 export { AbilityItemPF2e } from "./ability/document.ts";
+export { CampaignFeaturePF2e } from "./campaign-feature/document.ts";
 export { ConsumablePF2e } from "./consumable/document.ts";
 export { DeityPF2e } from "./deity/document.ts";
 export { FeatPF2e } from "./feat/document.ts";
 export { HeritagePF2e } from "./heritage/document.ts";
-export { CampaignFeaturePF2e } from "./campaign-feature/document.ts";
 export { KitPF2e } from "./kit/document.ts";
-export { LorePF2e } from "./lore/index.ts";
+export { LorePF2e } from "./lore.ts";
 export { MeleePF2e } from "./melee/document.ts";
 export { ShieldPF2e } from "./shield/document.ts";
 export { SpellPF2e } from "./spell/document.ts";
