+++ static/templates/items/lore-details.hbs
@@ -1,16 +1,33 @@
-<table class="skill-variants">
-    <caption>Situational Skill Rolls</caption>
-    <tr>
-        <th>Label</th>
-        <th>Roll Options</th>
-        <th>&nbsp;</th>
-    </tr>
-    {{#each data.variants as |variant idx|}}
-    <tr>
-        <td><input name="system.variants.{{idx}}.label" value="{{variant.label}}"></td>
-        <td><input name="system.variants.{{idx}}.options" value="{{variant.options}}" placeholder="separated by comma"></td>
-        <td><div class="remove-skill-variant" data-skill-variant-index="{{idx}}" title="{{localize "PF2E.ItemRemoveRuleElement"}}"><i class="fas fa-trash"></i></div></td>
-    </tr>
-    {{/each}}
-</table>
-<div class="add-skill-variant"><i class="fas fa-plus"></i> {{localize "PF2E.ItemNewSkillVariant"}}</div>
+<fieldset class="basics">
+    <div class="form-group">
+        <label>{{localize "PF2E.ProficiencyLabel"}}</label>
+        <select name="system.proficient.value" data-dtype="Number">
+            {{#select data.proficient.value}}
+                {{#each proficiencies as |label p|}}
+                    <option value="{{p}}">{{localize label}}</option>
+                {{/each}}
+            {{/select}}
+        </select>
+    </div>
+</fieldset>
+
+<fieldset>
+    <legend>Situational Skill Rolls</legend>
+
+    <table class="skill-variants">
+        <tr>
+            <th>Label</th>
+            <th>Roll Options</th>
+            <th>&nbsp;</th>
+        </tr>
+        {{#each data.variants as |variant idx|}}
+        <tr>
+            <td><input name="system.variants.{{idx}}.label" value="{{variant.label}}"></td>
+            <td><input name="system.variants.{{idx}}.options" value="{{variant.options}}" placeholder="separated by comma"></td>
+            <td><div class="remove-skill-variant" data-skill-variant-index="{{idx}}" title="{{localize "PF2E.ItemRemoveRuleElement"}}"><i class="fa-solid fa-trash"></i></div></td>
+        </tr>
+        {{/each}}
+    </table>
+    <div class="add-skill-variant"><i class="fa-solid fa-plus"></i> {{localize "PF2E.ItemNewSkillVariant"}}</div>
+    <div class="form-group">{{localize "PF2E.LoreProficiencyDisclaimer"}}</div>
+</fieldset>
