--- static/templates/items/lore-sidebar.hbs
@@ -1,13 +1,0 @@
-<div class="inventory-details">
-    <div class="form-group">
-        <label>{{localize "PF2E.ProficiencyLabel"}}</label>
-        <select name="system.proficient.value" data-dtype="Number">
-            {{#select data.proficient.value}}
-                {{#each proficiencies as |label p|}}
-                    <option value="{{p}}">{{localize label}}</option>
-                {{/each}}
-            {{/select}}
-        </select>
-    </div>
-    <div class="form-group">{{localize "PF2E.LoreProficiencyDisclaimer"}}</div>
-</div>
